+++ src/app/cart/cart.module.ts
@@ -2,8 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { SharedModule } from '../shared/shared.module';
-import { CartComponent } from './components/cart-list/cart-list.component';
-import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { CartItemComponent, CartListComponent} from './components';
 
 @NgModule({
   imports: [
@@ -12,11 +11,11 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
     FormsModule
   ],
   declarations: [
-    CartComponent,
+    CartListComponent,
     CartItemComponent
   ],
   exports: [
-    CartComponent
+    CartListComponent
   ]
 })
 export class CartModule {