+++ src/app/core/services/index.ts
@@ -1,0 +1,4 @@
+export * from './local-storage.service';
+export * from './config-options.service';
+export * from './constants.service';
+export * from './generator.service';