+++ src/app/products/components/index.ts
@@ -1,0 +1,2 @@
+export * from './product/product.component';
+export * from './products-list/products-list.component';