+++ src/app/products/products.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ProductsListComponent } from './components/products-list/products-list.component';
-import { ProductComponent } from './components/product/product.component';
+import { ProductsListComponent, ProductComponent } from './components';
 
 @NgModule({
   imports: [