+++ src/app/shared/models/config.model.ts
@@ -1,0 +1,8 @@
+export class ConfigModel {
+  constructor(public id,
+              public firstName: string,
+              public lastName: string,
+              public email: string,
+              public lastUpdateTime: Date) {
+  }
+}