+++ src/app/shared/models/interfaces/cart-map.interface.ts
@@ -1,0 +1,5 @@
+import { CartItemInterface } from './cart.interface';
+
+export interface CartMapInterface {
+  [name: string]: CartItemInterface;
+}