+++ src/app/shared/services/cart.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Product } from '../models/product.model';
-import { CartItemModel } from '../models/cart-item.model';
+
+import { CartItemModel, Product } from '../models';
+import { CartMapInterface } from '../models/interfaces/cart-map.interface';
+
 import { Subject } from 'rxjs';
 
 @Injectable({
@@ -8,49 +10,54 @@ import { Subject } from 'rxjs';
 })
 export class CartService {
 
-  public total: number;
-  public cart: Object;
+  public totalPrice = 0;
+  public totalCount = 0;
+  public cart: CartMapInterface = {};
   private channel = new Subject<Object>();
   public channel$ = this.channel.asObservable();
-  private totalChannel = new Subject<number>();
+  private totalChannel = new Subject<{ price: number, count: number }>();
   public totalChannel$ = this.totalChannel.asObservable();
 
   constructor() {
-    this.cart = [];
-    this.total = 0;
+
   }
 
-  putToCart(item: Product) {
+  putToCart(item: Product, quantity: number = 1) {
     this.cart[item.name] = this.cart[item.name] || new CartItemModel(item.id, item.name, 0, item.price);
-    this.cart[item.name]['quantity']++;
-    console.log(item.price);
-    this.total += item.price;
+    this.cart[item.name]['quantity'] += quantity;
+    this.increaseTotals(item.price, quantity);
     this.notify();
   }
 
-  removeFromCart(item) {
+  removeFromCart(item: Product, quantity: number = 1) {
     if (this.cart[item.name]['quantity'] === 1) {
       delete this.cart[item.name];
+      this.decreaseTotals(item.price, 1);
     } else {
-      this.cart[item.name]['quantity']--;
+      this.cart[item.name]['quantity'] -= quantity;
+      this.decreaseTotals(item.price, quantity);
     }
-    this.total -= item.price;
     this.notify();
   }
 
   updateQuantity(item) {
     this.cart[item.name]['quantity'] = item.quantity;
-    this.total = this.getTotal();
+    const keys = Object.keys(this.cart);
+    for (const key of keys) {
+      this.totalPrice += this.cart[key]['price'] * this.cart[key]['quantity'];
+      this.totalCount += this.cart[key]['quantity'];
+    }
     this.notify();
   }
 
-  getTotal(): number {
-    const keys = Object.keys(this.cart);
-    let total = 0;
-    for (const key of keys) {
-      total += this.cart[key]['price'] * this.cart[key]['quantity'];
-    }
-    return total;
+  increaseTotals(price: number, quantity: number) {
+    this.totalPrice += (price * quantity);
+    this.totalCount += quantity;
+  }
+
+  decreaseTotals(price: number, quantity: number) {
+    this.totalPrice -= (price * quantity);
+    this.totalCount -= quantity;
   }
 
   getCart() {
@@ -59,12 +66,13 @@ export class CartService {
 
   notify() {
     this.channel.next(this.getCart());
-    this.totalChannel.next(this.total);
+    this.totalChannel.next({price: this.totalPrice, count: this.totalCount});
   }
 
   clear() {
-    this.cart = [];
-    this.total = 0;
+    this.cart = {};
+    this.totalPrice = 0;
+    this.totalCount = 0;
     this.notify();
   }
 }